backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.26.0
anthropic>=0.40.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
xlrd>=2.0.1
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
